git-svn-id: https://localhost/svn/Project64/trunk@41 111125ac-702d-7242-af9c-5ba8ae61c1ef

+++ Source/Project64/N64 System/Mips/Audio.h
@@ -20,7 +20,7 @@ private:
 	DWORD  m_SecondBuff;
 	DWORD  m_Status;
 	DWORD  m_BytesPerSecond;
-	double m_CountsPerByte;
+	int    m_CountsPerByte;
 	int    m_FramesPerSecond;
 
 /*	float  m_VSyncTiming;
